+++ src/main/java/net/dzikoysk/funnyguilds/system/validity/ValiditySystem.java
@@ -2,6 +2,9 @@ package net.dzikoysk.funnyguilds.system.validity;
 
 import net.dzikoysk.funnyguilds.basic.Guild;
 import net.dzikoysk.funnyguilds.basic.util.GuildUtils;
+import net.dzikoysk.funnyguilds.event.SimpleEventHandler;
+import net.dzikoysk.funnyguilds.event.FunnyEvent.EventCause;
+import net.dzikoysk.funnyguilds.event.guild.GuildDeleteEvent;
 
 public class ValiditySystem {
 
@@ -15,15 +18,21 @@ public class ValiditySystem {
         if (instance == null) {
             new ValiditySystem();
         }
+        
         return instance;
     }
 
     public void run() {
         for (Guild guild : GuildUtils.getGuilds()) {
             if (!guild.isValid()) {
+                if (!SimpleEventHandler.handle(new GuildDeleteEvent(EventCause.SYSTEM, null, guild))) {
+                    continue;
+                }
+                
                 ValidityUtils.broadcast(guild);
                 GuildUtils.deleteGuild(guild);
             }
         }
     }
+    
 }
